test/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-appconfiguration==1.4.0
azure-appconfiguration-provider==1.0.0
azure-common==1.1.28
azure-core==1.29.4
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.14.1
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.0.0
azure-monitor-opentelemetry-exporter==1.0.0b17
blinker==1.6.2
boto3==1.28.57
botocore==1.31.57
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==4.4.2
Deprecated==1.2.14
exceptiongroup==1.1.3
fixedint==0.1.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
moviepy==1.0.3
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
opencv-python-headless==4.8.1.78
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-instrumentation-dbapi==0.41b0
opentelemetry-instrumentation-django==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-psycopg2==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-urllib==0.41b0
opentelemetry-instrumentation-urllib3==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
opentelemetry-resource-detector-azure==0.1.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
outcome==1.2.0
packaging==23.2
pdf2image==1.16.3
Pillow==10.0.0
portalocker==2.8.2
proglog==0.1.10
protobuf==4.24.2
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyee==8.2.2
pygifsicle==1.0.7
PyJWT==2.8.0
PyMuPDF==1.23.2
PyMuPDFb==1.23.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
pytube==15.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
selenium==4.15.2
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starkbank-ecdsa==2.2.0
support-developer==1.0.5
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
webdriver-manager==4.0.1
websockets==10.4
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
